requirements.txt
numpy
packaging
pydra_config
torch
tqdm
datasets
transformers
pytest
ninja
accelerate
bitsandbytes
wandb
trl
peft
vllm
flash_attn
deepspeed